componentized user open orders, history, and settled balance
